Ignore forwards from the bot itself
